Move as many types from core to format as possible
