+++ crates/hayagriva-format/Cargo.toml
@@ -21,10 +21,7 @@ unicode-segmentation = { workspace = true }
 hayagriva-core = { workspace = true }
 thiserror = { workspace = true }
 unscanny = { workspace = true }
-url = { workspace = true, optional = true }
+url = { workspace = true }
 
 [dev-dependencies]
 serde_json = "1"
-
-[features]
-biblatex = ["dep:biblatex", "dep:url"]
